requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
google-generativeai==0.3.2
reportlab==4.0.7
openpyxl==3.1.2